Add browserify babelify option to package.json
This change allows the babelify transform to be used in a browserify project (in my case, a browser extension project)

Adding `.es6.js` to `extensions` must be done on the main browserify instance (like, in the project's gulpfile). It doesn't seem to obey it if I only have it in Snoode's package.json for some reason.