Add app=mweb-{env} to API calls
Add an app name querystring so that it's easy to parse out when reading
logs on the API side. Preferably this will happen eventually by doing
something like parsing JWTs from oauth tokens and adding a tracer on
the layer above the API, but that's likely a long ways off. This'll do
for now.