Normalize Messages (this one will need more thought, see git log)
There might have to be fancy merge’ing of data from the messages
endpoint. You get back comments that are sent to your inbox. This
happens when ‘send replies to my inbox’ is on, and its on by default.
The comment that comes back doesn’t have score, or other parts to the
comment, so be careful about how it’s used as an update if that comment
is already in your data store