Make subreddit's readable short-name the uuid. Alias key_color
By making the short name the uuid we can look up the subreddit in our
cache without keeping an index for the short name -> thing id. This
seems like the more common lookup scenario as the subreddit name is in
the url for: 1) listings by subreddit e.g. /r/javascript and 2)
comments /r/javascript/comments/1234/stuff-about-javascript