Expose the ids for pagination to api consumers.
Subreddit’s uuid is the ‘url name’ because of how they’re parsed and
looked up in most clients. This makes paging difficult — there were
workarounds internally to api-client. This causes problems when
api-client consumers want to do their own subreddit pagination with
Records directly. By exposing the id of records for use in pagination,
this problem is fixed.