Stop posting endpoint from throwing invalid exceptions
The posting endpoint wrapper was throwing exceptions when
json errors were present (not 4xx's/5xx's but a 200). Since
we actually need to look at the errors stop this from
happening.

I handled this by not having a default `parseBody` function
but allowing the mobile code to handle the promise more
directly.