Make checking in and publishing code simpler/more correct
Problem:
Currently we check-in our transpiled code. Other than this being a
general nuisance, it's actually a little dangerous as well. Because it's
essentially one big blob of code, it's hard to read and parse. We've
already encountered situations when someone's checked in code bashed
someone else's changes.

Fix:
It's fairly simple to fix this by checking in only the source code and
making sure we publish the build code.

Details:
I've moved all our source code into /src, that way it's easier to target
for things like our linter. This change is actually straightforward
since all the paths are relative and therefore I only needed to update
webpack's config to point to /src/index.es6.js.

I then updated package.json to have a prepublish hook to ensure that
our source code is transpiled before it's pushed to npm.

Other changes:

- Removed a bunch of gitignore things that looked irrelevant to this
  library
- Removed .editorconfig and mancyLoader, which also looks irrelevant.
- Updated package.json with more up to date information and updated
  build to 0.10.0 (tested this out as well).
- Added a bunch of stuff that shouldn't get published to npmignore.

There's some other funky stuff going on in this repo, such as js files
being executable and the .es6.js extensions but in the interest of
smaller changes (and this is already a pretty big one), I'm not going to
touch those on this pass.