Fix babel error with super and writing readonly props
There are two big fixes in this patch, the first is a
simple issue with readonly props. FakeError has readonly
`message` and `stack` properties that were overwitten with
calls to `Object.assign`. The fix is an alternative helper
that won't overrwrite those props.

The second fix is for a bug in babel. The ResponseError
constructor tries to bail out and return a different
FakeError subclass to differentiate Disconnection Errors.
Babel was inserting its guard function to ensure super was
called before the constructor exits (which we don't need
because we bail out explicitly).