fixup - check unaliased and aliased prop when creating uuid
The raw data is passed into makeUUID, before mapping the real
api keys to their aliases.  This means that when a model is
instantiated from data from the api, we need to check the
unmapped key, but if the model is instantiated from a local
update, we need to check the aliased key.  This should probably
be fixed in general, but to avoid scope creep I'm just going
to fix it here.